+++ examples/cognite_core/_api_client.py
@@ -50,7 +50,7 @@ class CogniteCoreClient:
     CogniteCoreClient
 
     Generated with:
-        pygen = 0.99.50
+        pygen = 0.99.51
         cognite-sdk = 7.66.0
         pydantic = 2.9.2
 
@@ -68,7 +68,7 @@ class CogniteCoreClient:
         else:
             raise ValueError(f"Expected CogniteClient or ClientConfig, got {type(config_or_client)}")
         # The client name is used for aggregated logging of Pygen Usage
-        client.config.client_name = "CognitePygen:0.99.50"
+        client.config.client_name = "CognitePygen:0.99.51"
 
         self._client = client