+++ examples/omni/data_classes/_connection_item_b.py
@@ -120,9 +120,9 @@ class ConnectionItemBGraphQL(GraphQLCore):
                 last_updated_time=self.data_record.last_updated_time,
                 created_time=self.data_record.created_time,
             ),
-            inwards=[inward.as_read() for inward in self.inwards or []],
+            inwards=[inward.as_read() for inward in self.inwards] if self.inwards is not None else None,
             name=self.name,
-            self_edge=[self_edge.as_read() for self_edge in self.self_edge or []],
+            self_edge=[self_edge.as_read() for self_edge in self.self_edge] if self.self_edge is not None else None,
         )
 
     # We do the ignore argument type as we let pydantic handle the type checking
@@ -133,9 +133,9 @@ class ConnectionItemBGraphQL(GraphQLCore):
             space=self.space,
             external_id=self.external_id,
             data_record=DataRecordWrite(existing_version=0),
-            inwards=[inward.as_write() for inward in self.inwards or []],
+            inwards=[inward.as_write() for inward in self.inwards] if self.inwards is not None else None,
             name=self.name,
-            self_edge=[self_edge.as_write() for self_edge in self.self_edge or []],
+            self_edge=[self_edge.as_write() for self_edge in self.self_edge] if self.self_edge is not None else None,
         )
 
 
@@ -171,12 +171,20 @@ class ConnectionItemB(DomainModel):
             space=self.space,
             external_id=self.external_id,
             data_record=DataRecordWrite(existing_version=self.data_record.version),
-            inwards=[inward.as_write() if isinstance(inward, DomainModel) else inward for inward in self.inwards or []],
+            inwards=(
+                [inward.as_write() if isinstance(inward, DomainModel) else inward for inward in self.inwards]
+                if self.inwards is not None
+                else None
+            ),
             name=self.name,
-            self_edge=[
-                self_edge.as_write() if isinstance(self_edge, DomainModel) else self_edge
-                for self_edge in self.self_edge or []
-            ],
+            self_edge=(
+                [
+                    self_edge.as_write() if isinstance(self_edge, DomainModel) else self_edge
+                    for self_edge in self.self_edge
+                ]
+                if self.self_edge is not None
+                else None
+            ),
         )
 
     def as_apply(self) -> ConnectionItemBWrite: