+++ examples/omni_sub/data_classes/_connection_item_a.py
@@ -132,7 +132,7 @@ class ConnectionItemAGraphQL(GraphQLCore):
             other_direct=(
                 self.other_direct.as_read() if isinstance(self.other_direct, GraphQLCore) else self.other_direct
             ),
-            outwards=[outward.as_read() for outward in self.outwards or []],
+            outwards=[outward.as_read() for outward in self.outwards] if self.outwards is not None else None,
             self_direct=self.self_direct.as_read() if isinstance(self.self_direct, GraphQLCore) else self.self_direct,
         )
 
@@ -148,7 +148,7 @@ class ConnectionItemAGraphQL(GraphQLCore):
             other_direct=(
                 self.other_direct.as_write() if isinstance(self.other_direct, GraphQLCore) else self.other_direct
             ),
-            outwards=[outward.as_write() for outward in self.outwards or []],
+            outwards=[outward.as_write() for outward in self.outwards] if self.outwards is not None else None,
             self_direct=self.self_direct.as_write() if isinstance(self.self_direct, GraphQLCore) else self.self_direct,
         )
 
@@ -189,9 +189,11 @@ class ConnectionItemA(DomainModel):
             other_direct=(
                 self.other_direct.as_write() if isinstance(self.other_direct, DomainModel) else self.other_direct
             ),
-            outwards=[
-                outward.as_write() if isinstance(outward, DomainModel) else outward for outward in self.outwards or []
-            ],
+            outwards=(
+                [outward.as_write() if isinstance(outward, DomainModel) else outward for outward in self.outwards]
+                if self.outwards is not None
+                else None
+            ),
             self_direct=self.self_direct.as_write() if isinstance(self.self_direct, DomainModel) else self.self_direct,
         )